requirements.txt
Flask
firebase_admin
google.cloud
gunicorn
google-cloud-secret-manager